Detect short reads, bound callback concurrency, and release at free
Addresses review feedback on the C FFI callback reader.

The read callback now returns the number of bytes written rather than a
status code, so a host that reports success without filling the buffer is
rejected instead of handing uninitialized memory to the scan. The Java
bindings already guard this via `remaining()`; the previous C signature made
it undetectable.

A source was owned by the segment source's spawned read driver, which the
FFI's current-thread runtime never polled again once a call returned, so
`release` ran at an unpredictable later point or not at all. Hosts freeing
their context after `vx_data_source_free` could therefore hit a
use-after-free. `CurrentThreadRuntime::drain` runs queued tasks to a stop,
and the FFI drains after dropping a data source and after a failed open, so
`release` now runs before the call that drops the source returns.

Callback concurrency is additionally capped process-wide rather than only
per source, matching how the Java bindings bound upcalls across a
filesystem, and a name view of NULL with a non-zero length is rejected
rather than silently treated as anonymous.

+++ vortex-ffi/cinclude/vortex.h
@@ -561,8 +561,8 @@ typedef struct {
      */
     uint64_t len;
     /**
-     * Maximum number of concurrent "read_at" calls. 0 selects a default. The
-     * cap is per-source, so opening many files multiplies it.
+     * Maximum number of concurrent "read_at" calls for this source. 0 selects a
+     * default. A process-wide ceiling applies across all sources as well.
      */
     size_t concurrency;
     /**
@@ -571,13 +571,13 @@ typedef struct {
      */
     vx_view name;
     /**
-     * Required. Must write all "length" bytes at "offset" into "dst" and return
-     * 0, or return non-zero; success without filling "dst" leaks uninitialized
-     * memory into the scan.
+     * Required. Writes "length" bytes at "offset" into "dst" and returns the
+     * count written; a short count or a negative value fails the read.
      */
-    int32_t (*read_at)(void *ctx, uint64_t offset, uint8_t *dst, size_t length);
+    int64_t (*read_at)(void *ctx, uint64_t offset, uint8_t *dst, size_t length);
     /**
-     * Optional. Called once, after Vortex has dropped the source.
+     * Optional. Called once, before the call that drops the source returns -
+     * on that thread, or on a worker thread if any are configured.
      */
     void (*release)(void *ctx);
 } vx_readat;
@@ -937,7 +937,8 @@ vx_data_source_new_buffer(const vx_session *session, const void *buffer, size_t 
  *
  * "reader" is read during this call only; its callbacks and context must stay
  * valid until "release" runs. A rejected descriptor leaves ownership with the
- * caller and never calls "release"; once accepted, "release" always runs.
+ * caller and never calls "release"; once accepted, "release" runs before this
+ * call returns if it fails, and otherwise before vx_data_source_free returns.
  *
  * On error, returns NULL and sets "err".
  */
